+++ lib/redis_failover/node.rb
@@ -118,7 +118,6 @@ module RedisFailover
     end
     alias_method :eql?, :==
 
-
     # @return [Integer] a hash value for this node
     def hash
       to_s.hash
@@ -175,14 +174,14 @@ module RedisFailover
         redis = new_client
         yield redis
       end
-    rescue
-      raise NodeUnavailableError, self, caller
+    rescue Exception => ex
+      raise NodeUnavailableError, "#{ex.class}: #{ex.message}", ex.backtrace
     ensure
       if redis
         begin
           redis.client.disconnect
-        rescue
-          raise NodeUnavailableError, self, caller
+        rescue Exception => ex
+          raise NodeUnavailableError, "#{ex.class}: #{ex.message}", ex.backtrace
         end
       end
     end